WebM (VP8) recorder and Jitter Buffer refactoring - refactored VP8 frame builder and resolved issue with artifacts - refactored Jitter Buffer usage - mirrored ebml-go package (WebM) and patched it to get rid of null termination in encoder

+++ internal/webrtc/utils/jitter_buffer.go
@@ -11,7 +11,7 @@ type JitterBuffer struct {
 	size      int64
 }
 
-func NewJitterBuffer(size int) *JitterBuffer {
+func NewJitterBuffer(size uint16) *JitterBuffer {
 	return &JitterBuffer{
 		packets: make([]*rtp.Packet, size),
 		size:    int64(size),
